Remove filingDataDto and rework tests to account for this. DataDto not desired as the format does not fit that used by software filers
