+++ src/test/java/uk/gov/companieshouse/officerfiling/api/controller/OfficerFilingControllerImplIT.java
@@ -23,7 +23,6 @@ import uk.gov.companieshouse.api.model.transaction.Transaction;
 import uk.gov.companieshouse.api.model.transaction.TransactionStatus;
 import uk.gov.companieshouse.api.sdk.ApiClientService;
 import uk.gov.companieshouse.logging.Logger;
-import uk.gov.companieshouse.officerfiling.api.model.dto.OfficerFilingDataDto;
 import uk.gov.companieshouse.officerfiling.api.model.dto.OfficerFilingDto;
 import uk.gov.companieshouse.officerfiling.api.model.entity.OfficerFiling;
 import uk.gov.companieshouse.officerfiling.api.model.entity.OfficerFilingData;
@@ -160,12 +159,10 @@ class OfficerFilingControllerImplIT {
     @Test
     void createFilingWhenTM01PayloadOKThenResponse201() throws Exception {
         final var body = "{" + TM01_FRAGMENT + "}";
-        final var dtoData = OfficerFilingDataDto.builder().referenceEtag("etag")
+        final var dto = OfficerFilingDto.builder()
+                .referenceEtag("etag")
                 .referenceAppointmentId(FILING_ID)
                 .resignedOn(LocalDate.of(2022, 9, 13))
-                .build();
-        final var dto = OfficerFilingDto.builder()
-            .data(dtoData)
             .build();
         var offData = new OfficerFilingData(
                 "etag",
@@ -201,12 +198,10 @@ class OfficerFilingControllerImplIT {
     @Test
     void patchFilingWhenFullTM01PayloadOKThenResponse200() throws Exception {
         final var body = "{" + TM01_FRAGMENT + "}";
-        final var dtoData = OfficerFilingDataDto.builder().referenceEtag("etag")
+        final var dto = OfficerFilingDto.builder()
+                .referenceEtag("etag")
                 .referenceAppointmentId(FILING_ID)
                 .resignedOn(LocalDate.of(2022, 9, 13))
-                .build();
-        final var dto = OfficerFilingDto.builder()
-                .data(dtoData)
                 .build();
         var offData = new OfficerFilingData(
                 "etag",
@@ -242,11 +237,9 @@ class OfficerFilingControllerImplIT {
     @Test
     void patchFilingWhenMissingResignedOnTM01PayloadOKThenResponse200() throws Exception {
         final var body = "{" + PARTIAL_TM01_FRAGMENT_MISSING_RESIGNED_ON + "}";
-        final var dtoData = OfficerFilingDataDto.builder().referenceEtag("etag")
-                .referenceAppointmentId(FILING_ID)
-                .build();
         final var dto = OfficerFilingDto.builder()
-                .data(dtoData)
+                .referenceEtag("etag")
+                .referenceAppointmentId(FILING_ID)
                 .build();
         var offData = new OfficerFilingData(
                 "etag",
@@ -282,11 +275,9 @@ class OfficerFilingControllerImplIT {
     @Test
     void patchFilingWhenMissingEtagTM01PayloadOKThenResponse200() throws Exception {
         final var body = "{" + PARTIAL_TM01_FRAGMENT_MISSING_ETAG + "}";
-        final var dtoData = OfficerFilingDataDto.builder().referenceAppointmentId(FILING_ID)
-                .resignedOn(LocalDate.of(2022, 9, 13))
-                .build();
         final var dto = OfficerFilingDto.builder()
-                .data(dtoData)
+                .referenceAppointmentId(FILING_ID)
+                .resignedOn(LocalDate.of(2022, 9, 13))
                 .build();
         var offData = new OfficerFilingData(
                 "",
@@ -362,11 +353,9 @@ class OfficerFilingControllerImplIT {
         final var now = clock.instant();
         final var filing = OfficerFiling.builder().createdAt(now).updatedAt(now).data(offData)
                 .build();
-        final var dtoData = OfficerFilingDataDto.builder().referenceEtag("etag")
-                .referenceAppointmentId(FILING_ID)
-                .build();
         final var dto = OfficerFilingDto.builder()
-                .data(dtoData)
+                .referenceEtag("etag")
+                .referenceAppointmentId(FILING_ID)
                 .build();
         when(clock.instant()).thenReturn(FIRST_INSTANT);
         when(officerFilingService.save(any(OfficerFiling.class), eq(TRANS_ID))).thenReturn(
@@ -412,12 +401,10 @@ class OfficerFilingControllerImplIT {
 
     @Test
     void getFilingForReviewThenResponse200() throws Exception {
-        final var dtoData = OfficerFilingDataDto.builder().referenceEtag("etag")
+        final var dto = OfficerFilingDto.builder()
+                .referenceEtag("etag")
                 .referenceAppointmentId(FILING_ID)
                 .resignedOn(LocalDate.of(2022, 9, 13))
-                .build();
-        final var dto = OfficerFilingDto.builder()
-                .data(dtoData)
                 .build();
         var offData = new OfficerFilingData(
                 "etag",
@@ -442,12 +429,10 @@ class OfficerFilingControllerImplIT {
 
     @Test
     void getFilingForReviewOnLeapYearThenResponse200() throws Exception {
-        final var dtoData = OfficerFilingDataDto.builder().referenceEtag("etag")
+        final var dto = OfficerFilingDto.builder()
+                .referenceEtag("etag")
                 .referenceAppointmentId("id")
                 .resignedOn(LocalDate.of(2024, 2, 29))
-                .build();
-        final var dto = OfficerFilingDto.builder()
-                .data(dtoData)
                 .build();
         var offData = new OfficerFilingData(
                 "etag",
@@ -485,12 +470,10 @@ class OfficerFilingControllerImplIT {
         final Instant resignedToday = Instant.now();
         String formattedDate = new SimpleDateFormat("yyyy-MM-dd").format(Date.from(resignedToday));
         final var body = "{" + TM01_FRAGMENT.replace("2022-09-13", formattedDate) + "}";
-        final var dtoData = OfficerFilingDataDto.builder().referenceEtag("etag")
+        final var dto = OfficerFilingDto.builder()
+                .referenceEtag("etag")
                 .referenceAppointmentId(FILING_ID)
                 .resignedOn(LocalDate.now())
-                .build();
-        final var dto = OfficerFilingDto.builder()
-                .data(dtoData)
                 .build();
         var offData = new OfficerFilingData(
                 "etag",
@@ -581,12 +564,10 @@ class OfficerFilingControllerImplIT {
     @Test
     void createFilingWhenTransactionIsNull() throws Exception {
         final var body = "{" + TM01_FRAGMENT + "}";
-        final var dtoData = OfficerFilingDataDto.builder().referenceEtag("etag")
+        final var dto = OfficerFilingDto.builder()
+                .referenceEtag("etag")
                 .referenceAppointmentId(FILING_ID)
                 .resignedOn(LocalDate.of(2022, 9, 13))
-                .build();
-        final var dto = OfficerFilingDto.builder()
-                .data(dtoData)
                 .build();
 
         when(apiClientService.getApiClient(PASSTHROUGH_HEADER)).thenReturn(apiClientMock);
